+++ src/pages/menu/PostingDetailPage.tsx
@@ -1,21 +1,13 @@
-import React, { useEffect, useRef, useState } from 'react';
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
+import React, { useEffect, useState } from 'react';
+import { useLocation, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import { Posting } from 'types/posting';
 import { getPostingDetail } from 'utils/menu/getPostingDetail';
-import { ReactComponent as EyeSVG } from '../../assets/icons/menu/eye.svg';
-import { ReactComponent as MoreSVG } from '../../assets/icons/menu/more.svg';
-import { ReactComponent as LikeSVG } from '../../assets/icons/menu/like.svg';
 import { postPostingViewCount } from 'utils/menu/postPostingViewCount';
-import { postPostingLike } from 'utils/menu/postPostingLike';
-import '../../../node_modules/react-quill-new/dist/quill.snow.css';
-import ReactQuill, { Quill } from 'react-quill-new';
 import { patchPosting } from 'utils/menu/patchPosting';
-import { postCommentReply } from 'utils/menu/postCommentReply';
-import { deletePosting } from 'utils/menu/deletePosting';
-import HeadComment from 'components/menu/common/HeadComment';
-import ReplyComment from 'components/menu/common/ReplyComment';
-import ReportModalPortal from 'components/portal/ReportModalPortal';
+import PostInfoTitle from './PostInfoTitle';
+import PostContent from './PostContent';
+import PostComment from './PostComment';
 
 export default function PostingDetailPage() {
   const location = useLocation();
@@ -31,34 +23,12 @@ export default function PostingDetailPage() {
   const [like, setLike] = useState<boolean>(false);
   const [likeCount, setLikeCount] = useState<number>(0);
 
-  const [more, setMore] = useState<boolean>(false);
-  const moreRef = useRef<HTMLDivElement>(null);
-
   const [editting, setEditting] = useState<boolean>(false);
   const [title, setTitle] = useState<string>('');
   const [content, setContent] = useState<string>('');
   const [price, setPrice] = useState<string>('');
 
-  const textarea = useRef<HTMLTextAreaElement>(null);
   const [comment, setComment] = useState<string>('');
-
-  const [openReport, setOpenReport] = useState<boolean>(false);
-  const [reportType, setReportType] = useState<string>('');
-  const [reportReason, setReportReason] = useState<string>('');
-
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    if (openReport) {
-      document.body.style.overflow = 'hidden'; // 스크롤 비활성화
-    } else {
-      document.body.style.overflow = 'auto'; // 스크롤 활성화
-    }
-
-    return () => {
-      document.body.style.overflow = 'auto'; // 컴포넌트 언마운트 시 스크롤 복원
-    };
-  }, [openReport]);
 
   useEffect(() => {
     getPostingDetail(postingId, setPosting).then((result) => {
@@ -70,81 +40,12 @@ export default function PostingDetailPage() {
       setContent(result.content);
     });
     postPostingViewCount(postingId);
-
-    const handleClickOutside = (event: MouseEvent) => {
-      if (moreRef.current && !moreRef.current.contains(event.target as Node)) {
-        setMore(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
   }, []);
-
-  const onclickLikeButton = () => {
-    postPostingLike(postingId, setLike);
-    if (like) {
-      setLikeCount((prev) => prev - 1);
-    } else setLikeCount((prev) => prev + 1);
-  };
-
-  const onChangeTextArea = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    if (textarea.current) {
-      textarea.current.style.height = 'auto';
-      textarea.current.style.height = textarea.current.scrollHeight + 'px';
-    }
-    setComment(e.target.value);
-  };
-
-  const onClickDelete = () => {
-    deletePosting(postingId).then(() => {
-      navigate(`/menu/${state.boardType}`, { replace: true });
-    });
-  };
 
   const onClickSave = () => {
     patchPosting(postingId, title, content, price).then(() =>
       window.location.reload()
     );
-  };
-
-  const onClickCancle = () => {
-    setEditting(false);
-    if (posting) {
-      setContent(posting.content);
-    }
-  };
-
-  const onClickSubmitComment = () => {
-    postCommentReply(postingId, comment).then((newComment) => {
-      setComments((prevComments) => [...prevComments, newComment]); // 새 댓글 추가
-      setComment('');
-    });
-    setComment('');
-  };
-
-  const onClickReportInMore = () => {
-    setOpenReport((prev) => !prev);
-    setMore(false);
-  };
-
-  const modules = {
-    toolbar: [
-      [{ header: [1, 2, 3, false] }], // 헤더
-      [{ align: [] }], // 정렬
-      ['bold', 'italic', 'underline', 'strike'], // 텍스트 포맷
-      [{ list: 'ordered' }, { list: 'bullet' }], // 리스트
-      ['blockquote', 'code-block'], // 인용구, 코드 블록
-      [{ color: [] }, { background: [] }], // 텍스트 색상, 배경색
-      ['link', 'image'], // 링크, 이미지, 비디오
-      ['clean'] // 포맷 초기화
-    ],
-    ImageResize: {
-      parchment: Quill.import('parchment'),
-      modules: ['Resize', 'DisplaySize']
-    }
   };
 
   if (!posting) {
@@ -156,179 +57,38 @@ export default function PostingDetailPage() {
         <p>{state.pageName}</p>
       </PageNameBox>
 
-      <PostHeader>
-        <PostTitle>
-          {editting ? (
-            <input
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-            />
-          ) : (
-            <span>{posting.title}</span>
-          )}
-        </PostTitle>
-
-        {!editting && (
-          <PostInfo>
-            <div>
-              <span>{posting.writer_nickname}</span>
-              <span>{posting.created_at.split('T')[0]}</span>
-              <span>
-                <EyeSVG></EyeSVG> {posting.view_count}
-              </span>
-            </div>
-            <MoreDiv ref={moreRef} onClick={() => setMore((prev) => !prev)}>
-              <span style={{ cursor: 'pointer' }}>
-                <MoreSVG></MoreSVG>
-              </span>
-              {more &&
-                (posting.is_mine ? (
-                  <div>
-                    <div onClick={() => setEditting(true)}>Edit</div>
-                    <div onClick={onClickDelete}>Delete</div>
-                  </div>
-                ) : (
-                  <div>
-                    <div onClick={onClickReportInMore}>Report</div>
-                  </div>
-                ))}
-            </MoreDiv>
-          </PostInfo>
-        )}
-      </PostHeader>
+      <PostInfoTitle
+        posting={posting}
+        editting={editting}
+        title={title}
+        setTitle={setTitle}
+        setEditting={setEditting}
+      ></PostInfoTitle>
 
       {!editting && <LineGradient></LineGradient>}
 
-      <PostContentBox>
-        <Content>
-          <CustomReactQuill
-            $editting={editting}
-            value={editting ? content : posting.content}
-            readOnly={!editting}
-            theme="snow"
-            modules={modules}
-            onChange={(value) => editting && setContent(value)}
-          />
-        </Content>
-        {!editting && (
-          <LikeWrapper $like={like}>
-            <div onClick={onclickLikeButton}>
-              <LikeSVG></LikeSVG>
-            </div>
-            <p>{likeCount}</p>
-          </LikeWrapper>
-        )}
-        {editting && (
-          <ButtonWrapper>
-            <Button onClick={onClickCancle}>Cancel</Button>
-            <Button onClick={onClickSave}>Save</Button>
-          </ButtonWrapper>
-        )}
-      </PostContentBox>
+      <PostContent
+        posting={posting}
+        editting={editting}
+        content={content}
+        setContent={setContent}
+        like={like}
+        setLike={setLike}
+        likeCount={likeCount}
+        setLikeCount={setLikeCount}
+        setEditting={setEditting}
+        onClickSave={onClickSave}
+      ></PostContent>
 
-      {!editting ? (
-        <>
-          <LineGradient></LineGradient>
-
-          <PostCommentBox>
-            <p>
-              Comment <span>{posting.comment_count}</span>
-            </p>
-            <>{console.log(posting.comments)}</>
-            {comments.map((comment, index) => (
-              <React.Fragment key={index}>
-                <HeadComment
-                  key={comment.id}
-                  comment={comment}
-                  postingId={postingId}
-                ></HeadComment>
-              </React.Fragment>
-            ))}
-            <div>
-              <textarea
-                ref={textarea}
-                value={comment}
-                onChange={onChangeTextArea}
-              />
-              <SubmitButton onClick={onClickSubmitComment}>Submit</SubmitButton>
-            </div>
-          </PostCommentBox>
-        </>
-      ) : (
-        ''
-      )}
-
-      {openReport && (
-        <ReportModalPortal>
-          <ReportModal onClick={() => setOpenReport(false)}>
-            <div onClick={(e) => e.stopPropagation()}>
-              <p>Report</p>
-              <div>
-                <div>Please select report type</div>
-                <RadioWrapper>
-                  <label>
-                    <input
-                      type="radio"
-                      name="report_type"
-                      value="SPAM"
-                      onChange={(e) => setReportType(e.target.value)}
-                    />
-                    <span>Spam</span>
-                  </label>
-                  <label>
-                    <input
-                      type="radio"
-                      name="report_type"
-                      value="ABUSE"
-                      onChange={(e) => setReportType(e.target.value)}
-                    />
-                    <span>Abuse</span>
-                  </label>
-                  <label>
-                    <input
-                      type="radio"
-                      name="report_type"
-                      value="ADULT"
-                      onChange={(e) => setReportType(e.target.value)}
-                    />
-                    <span>Adult</span>
-                  </label>
-                  <label>
-                    <input
-                      type="radio"
-                      name="report_type"
-                      value="ILLEGAL"
-                      onChange={(e) => setReportType(e.target.value)}
-                    />
-                    <span>Illegal</span>
-                  </label>
-                  <label>
-                    <input
-                      type="radio"
-                      name="report_type"
-                      value="OTHER"
-                      onChange={(e) => setReportType(e.target.value)}
-                    />
-                    <span>Other</span>
-                  </label>
-                </RadioWrapper>
-
-                <WriteReason>
-                  <p>Reason</p>
-                  <textarea
-                    value={reportReason}
-                    onChange={(e) => setReportReason(e.target.value)}
-                  ></textarea>
-                </WriteReason>
-              </div>
-
-              <ReportButtonWrapper>
-                <ReportButton>Submit</ReportButton>
-              </ReportButtonWrapper>
-            </div>
-          </ReportModal>
-        </ReportModalPortal>
+      {!editting && <LineGradient></LineGradient>}
+      {!editting && (
+        <PostComment
+          posting={posting}
+          comments={comments}
+          comment={comment}
+          setComment={setComment}
+          setComments={setComments}
+        ></PostComment>
       )}
     </PostingDetailContainer>
   );
@@ -349,274 +109,9 @@ const PageNameBox = styled.div`
   }
 `;
 
-const PostHeader = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  margin-top: 6rem;
-`;
-
-const PostTitle = styled.p`
-  font-size: 2.5rem;
-  font-weight: 500;
-
-  > input {
-    font-size: 2.5rem;
-    width: 100%;
-    height: 5rem;
-    padding: 1.3rem;
-  }
-`;
-
-const PostInfo = styled.div`
-  display: flex;
-  justify-content: space-between;
-  color: var(--secondary-text);
-
-  > div {
-    display: flex;
-    gap: 2rem;
-    > span {
-    font-size: 1.6rem;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-  }
-`;
-
-const MoreDiv = styled.div`
-  position: relative;
-
-  > div {
-    background: white;
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-
-    top: 2.5rem;
-    right: 0;
-    width: 5rem;
-
-    > div {
-      &:hover {
-        background: var(--main-gray);
-      }
-      border: 1px solid var(--border-color);
-      cursor: pointer;
-      display: flex;
-
-      align-items: center;
-      height: 100%;
-      padding: 0.75rem;
-    }
-  }
-`;
-
 const LineGradient = styled.div`
   margin-top: 5rem;
   height: 0.1rem;
   background-image: var(--line-gradient);
   transform: rotate(180deg);
-`;
-
-const PostContentBox = styled.div`
-  margin-top: 3.5rem;
-
-  font-size: 3rem;
-`;
-
-const Content = styled.div`
-  margin-bottom: 6rem;
-`;
-
-const LikeWrapper = styled.div<{ $like: boolean }>`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-
-  > div {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 7.5rem;
-    height: 7.5rem;
-    background: ${(props) =>
-      props.$like ? 'var(--vertical-gradient)' : '#979797'};
-    border-radius: 100%;
-  }
-
-  > p {
-    font-size: 1.6rem;
-    font-weight: 600;
-  }
-`;
-
-const PostCommentBox = styled.div`
-  margin-top: 3rem;
-  padding: 1.5rem 2.5rem;
-
-  > p {
-    margin-bottom: 1rem;
-    font-size: 1.5rem;
-    font-weight: 300;
-
-    > span {
-      color: var(--primary-color);
-      font-weight: 400;
-    }
-  }
-
-  > div {
-    display: flex;
-    > textarea {
-      overflow-y: hidden;
-      font-size: 1.6rem;
-      padding: 1rem;
-      width: 100%;
-      height: 8.9rem;
-      resize: none;
-    }
-  }
-`;
-
-const CustomReactQuill = styled(ReactQuill)<{ $editting: boolean }>`
-  & .ql-container {
-    ${(props) =>
-      props.$editting
-        ? 'background: white; min-height: 26rem;'
-        : 'border: none;'}
-  }
-
-  & .ql-toolbar {
-    ${(props) => (props.$editting ? 'background: white;' : 'display: none;')}
-  }
-
-  & .ql-editor strong {
-    font-weight: bold;
-  }
-
-  & .ql-editor em {
-    font-style: italic;
-  }
-`;
-
-const ButtonWrapper = styled.div`
-  margin-top: 0rem;
-  display: flex;
-  justify-content: end;
-  gap: 4rem;
-`;
-
-const Button = styled.button`
-  width: 12rem;
-  height: 5rem;
-  flex-shrink: 0;
-  border-radius: 5px;
-  border: 1px solid #d6d6d6;
-  background: var(
-    --vertical-gradient,
-    linear-gradient(180deg, #b21f7c 22.5%, #4c0d0f 100%)
-  );
-  color: white;
-  font-size: 1.6rem;
-  font-weight: 600;
-  cursor: pointer;
-`;
-
-const SubmitButton = styled.button`
-  color: white;
-  width: 10rem;
-  height: 9rem;
-  flex-shrink: 0;
-  border-radius: 5px;
-  border: 1px solid #d6d6d6;
-  background: var(
-    --vertical-gradient,
-    linear-gradient(180deg, #b21f7c 22.5%, #4c0d0f 100%)
-  );
-`;
-
-const ReportModal = styled.div`
-  position: fixed;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: rgba(0, 0, 0, 0.2);
-  width: 100vw;
-  height: 100vh;
-
-  > div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border-radius: 0.5rem;
-    width: 40rem;
-    height: 50rem;
-    padding: 3rem;
-    background: white;
-
-    > p {
-      font-family: var(--sub-font);
-      color: var(--primary-color);
-      font-size: 4rem;
-      font-weight: 600;
-      margin-bottom: 4rem;
-    }
-
-    > div {
-      width: 100%;
-      font-size: 2.5rem;
-      > div {
-        margin-bottom: 1rem;
-      }
-    }
-  }
-`;
-
-const RadioWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-
-  font-size: 1.5rem;
-  > label {
-    display: block;
-  }
-`;
-
-const WriteReason = styled.div`
-  > p {
-    margin-bottom: 1rem;
-  }
-  > textarea {
-    resize: none;
-    width: 100%;
-    height: 7rem;
-  }
-`;
-
-const ReportButtonWrapper = styled.div`
-  margin-top: 1.5rem;
-  display: flex;
-  justify-content: center;
-`;
-
-const ReportButton = styled.button`
-  width: 12rem;
-  height: 5rem;
-  flex-shrink: 0;
-  border-radius: 5px;
-  border: 1px solid #d6d6d6;
-  background: var(
-    --vertical-gradient,
-    linear-gradient(180deg, #b21f7c 22.5%, #4c0d0f 100%)
-  );
-  color: white;
-  font-size: 1.6rem;
-  font-weight: 600;
-  cursor: pointer;
 `;
